+++ src/croniter/tests/test_croniter.py
@@ -1285,6 +1285,11 @@ class CroniterRangeTest(base.TestCase):
         with self.assertRaises(StopIteration):
             next(iterable)
 
+    def test_isssue151(self):
+        """."""
+        self.assertTrue(
+            croniter.match("* * * * *", datetime(2019, 1, 14, 11, 0, 59, 999999)))
+
 
 if __name__ == '__main__':
     unittest.main()