+++ docs/CHANGES.rst
@@ -4,8 +4,8 @@ Changelog
 1.0.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
-
+- Fix match when datetime has microseconds
+  [kiorky]
 
 1.0.1 (2021-01-06)
 ------------------